requirements.txt
openai
python-dotenv
urllib3
pydantic
